requirements.txt
aiohttp==3.7.4.post0
aiomisc==14.0.7
configargparse==1.4.1
requests==2.25.1
python-telegram-bot
